Use argparse for argument parsing (#2)
WARNING: This makes the script not backward-compatible due to the -l
becoming -d and a space being added for each parameter.
NOTE: If you wish to add a parameter starting with a dash which is
already an option (e.g. -S for cflags), you must precede it with a
space (-c " -S") so argparse doesn't mistake it for a source dir.